SINGA-233 Some code cleaning
