+++ include/singa/io/integer.h
